Kevin WIP Admin User - Admin - Mobil - Tambah Mobil
Tambah file : admin-mobil.php , tambah-mobil.php

Tambah coding di : kumpulan.php , register.css , register.php (untuk Email dan Kota)

Kurang Desain dashboard admin-user dan admin-mobil

Untuk admin-masteruser ikut di admin.php

+++ kumpulan.php
@@ -12,14 +12,16 @@ if($action == "Register"){
     $namalengkap = $_POST["nama"];
     $phone = $_POST["phone"];
     $alamat = $_POST["alamat"];
+    $kota = $_POST['kota'];
+    $email = $_POST['email'];
   
     
-    if($username != "" && $password != "" && $confirmpass != "" && $phone != "" && $alamat != "" && $nik != ""){
+    if($username != "" && $password != "" && $confirmpass != "" && $phone != "" && $alamat != "" && $nik != "" &&$kota != "" && $email != ""){
       $result = false;
       if($password == $confirmpass){
-              $stmt = $pdo -> prepare("INSERT INTO PENYEWA (nik	, username, password, nama, no_telp,alamat) VALUES (?,?,?,?,?,?)");
+              $stmt = $pdo -> prepare("INSERT INTO PENYEWA (nik	, username, password, nama, no_telp,alamat,kota,email) VALUES (?,?,?,?,?,?,?,?)");
               $result = $stmt-> execute([
-              $nik,$username,$password,$namalengkap,$phone,$alamat
+              $nik,$username,$password,$namalengkap,$phone,$alamat,$kota,$email
           ]);
       }else{
           echo "<script> alert('Password dan Confirm password Tidak sama'); </script>";
@@ -35,7 +37,29 @@ if($action == "Register"){
          header("Location:register.php");
       
     }else{
-        echo "<script> alert('ada field yang belum diisi'); </script>";
+        echo "<script> alert('Ada field yang belum diisi'); </script>";
+    }
+}else if($action=="addMobil"){
+    $nama = $_POST['nama_mobil'];
+    $tahun = $_POST['tahun'];
+    $bahan_bakar = $_POST['bahan_bakar'];
+    $jenis = $_POST['jenis'];
+    $tarif = $_POST['tarif'];
+    $status = "Available";
+
+    if($nama != "" && $tahun!="" && $bahan_bakar != "" && $jenis != ""){
+        $result = false;
+        $stmt = $pdo -> prepare("INSERT INTO MOBIL (nama_mobil, tahun, bahan_bakar,jenis, tarif, status) VALUES (?,?,?,?,?,?)");
+        $result = $stmt -> execute([$nama,$tahun,$bahan_bakar,$jenis,$tarif,$status]);
+        if($result == true){
+            $_SESSION["message"] = "Berhasil add nih";
+        }
+        else{
+            $_SESSION["message"] = "Gagal add nih";
+        }
+        header("Location:admin-mobil.php");
+    }else{
+        echo "<script> alert('Ada field yang belum diisi'); </script>";
     }
 }
